BHA-225 | Convert Network ids from int to long | Jayanta

+++ src/main/scala/com/bharatsim/model/House.scala
@@ -2,7 +2,7 @@ package com.bharatsim.model
 
 import com.bharatsim.engine.models.Network
 
-case class House(id: Int) extends Network {
+case class House(id: Long) extends Network {
   addRelation[Person]("HOUSES")
 
   override def getContactProbability(): Double = 1
